project/services/project-auth-service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
PyMySQL
SQLAlchemy
python-jose
passlib[bcrypt]